Add new deploy workflow, create draft on PR merge
Slight adjustment to our release process. On merge to main, we no longer
automatically publish a new npm package, but rather draft a new github
release.

The draft release is set up with suggested release notes, and also the
correct tag to create on publish. Once the release is published, a new
deployment workflow is run automatically to create the npm package.

The deployment progress will be posted automatically in Whereby Slack
deployments channel.