Refactor stopOrResumeVideo, handle localstream event
Extract all rtc related side-effects of stopping or resuming video into
new functions. This allows the existing .stopOrResumeVideo() function to
still use this logic, but also allow us to trigger the same logic on the
newly introduced `stopresumevideo` event which can be dispatched on the
localStream.

The rationale behind this change is to allow our new browser-sdk to get
the same side-effects when toggling video, but without maintaining a
direct reference to the rtc manager itself.