Remove internal devDependencies
We are seeing some problems from our customers when installing our
browser-sdk package. This is because we rely on some internal private
repos for development.

Originially we thought this would not be an issue, but it seems like we
were wrong. For now, copy the eslint config into this repo. This will
most likely be short-lived, as we are planning to move this library into
our new monorepo setup.