media: Ensure all relevant issue detectors check mute state
Changes in e0711dbdcb9f7ae03aa9e1ce87963d8c491206e9 meant that issue
detectors were invoked even if audio or video is disabled for a client.
That fixed a separate issue, but is problematic for other detectors. To
fix this, we now decorate all enabled checks for issue detectors which
expect media to be enabled, ensuring they are disabled if no longer
relevant.